requirements.txt
minio==7.2.7
python-dotenv==1.0.0
pyyaml==6.0.1
pyspark==3.5.0